解决登录验证码看不清问题，完成消费组的编辑，完成消费组关联用户弹窗（功能暂未实现）

+++ EasyLife/src/com/easylife/service/CostGroupService.java
@@ -1,9 +1,8 @@
 package com.easylife.service;
 
-import java.util.Iterator;
+import java.io.Serializable;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 
 import javax.annotation.Resource;
 
@@ -29,6 +28,10 @@ public class CostGroupService {
 		return selectAll;
 	}
 	
+	public CostGroup getById(Long groupId){
+		return groupDao.getById(CostGroup.class, groupId);
+	}
+	
 	public Page<CostGroup> selectPage(int currentPage, int pageSize, Map<String, String> params){
 		StringBuffer baseHql = new StringBuffer("from CostGroup cg order by cg.createTime desc");
 		Page<CostGroup> page = new Page<CostGroup>();
@@ -40,26 +43,36 @@ public class CostGroupService {
 	}
 	
 	public void save(CostGroup group, List<GroupMember> selectedMembers){
-		groupDao.save(group);
+		Serializable groupId = groupDao.save(group);
+		for (GroupMember groupMember : selectedMembers) {
+			groupMember.setGroup_id((Long) groupId);
+			memberDao.save(groupMember);
+		}
 	}
 	
 	public void update(CostGroup group, List<GroupMember> selectedMembers){
 		CostGroup foundgroup = groupDao.getById(CostGroup.class, group.getId());
-		List<GroupMember> foundMember = memberDao.findByGroupId(foundgroup.getId());
-		for (GroupMember groupMember : foundMember) {
-			memberDao.delete(groupMember);
+		if(selectedMembers != null && selectedMembers.size() > 0){
+			List<GroupMember> foundMember = memberDao.findByGroupId(foundgroup.getId());
+			for (GroupMember groupMember : foundMember) {
+				memberDao.delete(groupMember);
+			}
+			
+			for (GroupMember groupMember : selectedMembers) {
+				groupMember.setGroup_id(foundgroup.getId());
+				memberDao.save(groupMember);
+			}
+			foundgroup.setGroupName(group.getGroupName());
+			foundgroup.setMark(group.getMark());
 		}
-		
-		for (GroupMember groupMember : selectedMembers) {
-			groupMember.setGroup(foundgroup);
-			memberDao.save(groupMember);
-		}
-		foundgroup.setGroupName(group.getGroupName());
-		foundgroup.setMark(group.getMark());
 		groupDao.update(foundgroup);
 	}
 	
 	public void delete(CostGroup group){
+		List<GroupMember> foundMember = memberDao.findByGroupId(group.getId());
+		for (GroupMember groupMember : foundMember) {
+			memberDao.delete(groupMember);
+		}
 		groupDao.delete(group);
 	}
 }
